requirements.txt
asgiref==3.5.2
bleach==5.0.1
Django==3.2.16
django-cleanup==6.0.0
django-js-asset==2.0.0
django-summernote==0.8.20.0
Pillow==9.3.0
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
webencodings==0.5.1
